fix: 친구 요청 취소에서 // sender = currentUser 인지 확인

+++ bookduck/src/main/java/com/mmc/bookduck/domain/friend/service/FriendRequestService.java
@@ -51,6 +51,13 @@ public class FriendRequestService {
     public void cancelFriendRequest(Long requestId) {
         FriendRequest request = friendRequestRepository.findById(requestId)
                 .orElseThrow(() -> new CustomException(ErrorCode.FRIEND_REQUEST_NOT_FOUND));
+
+        // sender = currentUser 인지 확인
+        User currentUser = userService.getCurrentUser();
+        if (!request.getSender().getUserId().equals(currentUser.getUserId())) {
+            throw new CustomException(ErrorCode.UNAUTHORIZED_REQUEST);
+        }
+
         friendRequestRepository.delete(request);
     }
 
@@ -81,10 +88,12 @@ public class FriendRequestService {
         FriendRequest request = friendRequestRepository.findById(requestId)
                 .orElseThrow(() -> new CustomException(ErrorCode.FRIEND_REQUEST_NOT_FOUND));
         User currentUser = userService.getCurrentUser();
+
         // receiver = currentUser 인지 확인
         if (!request.getReceiver().getUserId().equals(currentUser.getUserId())) {
             throw new CustomException(ErrorCode.UNAUTHORIZED_REQUEST);
         }
+
         request.setFriendRequestStatus(FriendRequestStatus.REJECTED);
         friendRequestRepository.save(request);
     }
